NMS-9288: Topology on Map is not properly displayed
Refactored the name of classes
Every Provider now makes what his name
tells should make
Used Linkd as prefix everywhere
