+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/main/java/org/opennms/features/topology/plugins/topo/linkd/internal/LinkdCdpDetail.java
@@ -29,51 +29,44 @@
 package org.opennms.features.topology.plugins.topo.linkd.internal;
 
 import org.opennms.features.topology.api.topo.Vertex;
+import org.opennms.netmgt.model.CdpLink;
 
-public abstract class LinkdDetail<K,L> {
+public class LinkdCdpDetail extends LinkdEdgeDetail<CdpLink,CdpLink> {
 
-    private final String m_id;
-    private final Vertex m_source;
-    private final K m_sourceLink;
-    private final Vertex m_target;
-    private final L m_targetLink;
-
-    public LinkdDetail(String id, Vertex source, K sourceLink, Vertex target, L targetLink){
-        m_id = id;
-        m_source = source;
-        m_sourceLink = sourceLink;
-        m_target = target;
-        m_targetLink = targetLink;
+    public LinkdCdpDetail(String id, Vertex source, CdpLink sourceLink, Vertex target, CdpLink targetLink) {
+        super(id, source, sourceLink, target, targetLink);
     }
 
-    public abstract int hashCode();
-
-    public abstract boolean equals(Object obj);
-
-    public abstract Integer getSourceIfIndex();
-
-    public abstract Integer getTargetIfIndex();
-
-    public abstract String getType();
-
-    public String getId() {
-        return m_id;
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((getSourceLink() == null) ? 0 : getSource().getNodeID().hashCode()) + ((getTargetLink() == null) ? 0 : getTarget().getNodeID().hashCode());
+        return result;
     }
 
-    public Vertex getSource() {
-        return m_source;
+    @Override
+    public boolean equals(Object obj) {
+        if(obj instanceof LinkdCdpDetail){
+            LinkdCdpDetail objDetail = (LinkdCdpDetail)obj;
+
+            return getId().equals(objDetail.getId());
+        } else  {
+            return false;
+        }
     }
 
-    public Vertex getTarget() {
-        return m_target;
+    @Override
+    public Integer getSourceIfIndex() {
+        return getSourceLink().getCdpCacheIfIndex();
     }
 
-    public K getSourceLink() {
-        return m_sourceLink;
+    @Override
+    public Integer getTargetIfIndex() {
+        return getTargetLink().getCdpCacheIfIndex();
     }
 
-    public L getTargetLink() {
-        return m_targetLink;
-    }
+    @Override
+    public String getType() { return "CDP"; }
 
 }
