+++ features/topology-map/plugins/org.opennms.features.topology.plugins.topo.linkd/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -50,7 +50,7 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
         init-method="start"
         destroy-method="stop" />
 
-    <bean id="topologyProvider" scope="singleton" class="org.opennms.features.topology.plugins.topo.linkd.internal.EnhancedLinkdTopologyProvider">
+    <bean id="topologyProvider" scope="singleton" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdTopologyProvider">
         <argument ref="metricRegistry" />
         <property name="transactionOperations" ref="transactionOperations"/>
         <property name="nodeDao" ref="nodeDao" />
@@ -74,40 +74,45 @@ xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
         <argument ref="topologyProvider"/>
     </bean>
 
-    <bean id="topologyACLProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.NodeACLVertexProvider">
+    <bean id="linkdACLProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdACLVertexProvider">
         <argument ref="topologyHopProvider" />
         <argument ref="nodeDao" />
     </bean>
 
-    <!-- Meta topology provider -->
-    <bean id="enhancedLinkdMetaTopologyProvider" class="org.opennms.features.topology.api.topo.SimpleMetaTopologyProvider">
+    <!-- Search topology provider -->
+    <bean id="linkdSearchProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdSearchProvider">
         <argument ref="topologyACLProvider" />
     </bean>
 
-    <service interface="org.opennms.features.topology.api.topo.MetaTopologyProvider" ref="enhancedLinkdMetaTopologyProvider">
+    <service interface="org.opennms.features.topology.api.topo.SearchProvider" ref="linkdSearchProvider" />
+
+    <!-- Meta topology provider -->
+    <bean id="linkdMetaTopologyProvider" class="org.opennms.features.topology.api.topo.SimpleMetaTopologyProvider">
+        <argument ref="linkdACLProvider" />
+    </bean>
+
+    <service interface="org.opennms.features.topology.api.topo.MetaTopologyProvider" ref="linkdMetaTopologyProvider">
         <service-properties>
             <entry key="label" value="Enhanced Linkd" />
         </service-properties>
     </service>
-
-    <service interface="org.opennms.features.topology.api.topo.SearchProvider" ref="topologyACLProvider" />
 
     <!-- EnhancedLinkd StatusProviders Service and CheckedOperation -->
     <bean id="linkdEdgeStatusProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdEdgeStatusProvider">
         <property name="alarmDao" ref="alarmDao"/>
     </bean>
 
-	<bean id="enlinkdWrappedEdgeStatusProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdWrappedEdgeStatusProviders" init-method="init">
+	<bean id="linkdWrappedEdgeStatusProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdWrappedEdgeStatusProviders" init-method="init">
         <property name="edgeStatusProvider" ref="linkdEdgeStatusProvider"/>
     </bean>
 
-    <service interface="org.opennms.features.topology.api.topo.EdgeStatusProvider" ref="enlinkdWrappedEdgeStatusProvider" />
+    <service interface="org.opennms.features.topology.api.topo.EdgeStatusProvider" ref="linkdWrappedEdgeStatusProvider" />
 
     <!-- status provider -->
-    <bean id="statusProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.AlarmStatusProvider">
+    <bean id="linkdStatusProvider" class="org.opennms.features.topology.plugins.topo.linkd.internal.LinkdStatusProvider">
         <argument ref="alarmDao"/>
     </bean>
-    <service interface="org.opennms.features.topology.api.topo.StatusProvider" ref="statusProvider" />
+    <service interface="org.opennms.features.topology.api.topo.StatusProvider" ref="linkdStatusProvider" />
 
     <!-- Icon definition -->
     <service>
